Refactor project loading and validation workflow
- Updated open project form to clarify validation process and remove quick validate button.
- Modified project page description to emphasize full validation checks in the Validator.
- Enhanced frontend rebuild plan documentation to reflect current progress and remaining tasks.
- Added new JavaScript modules for project preflight checks, page bootstrap, project selection, and recent projects management.
- Implemented tests for recent projects controller and project selection functionality.
- Ensured separation of concerns in project loading and validation logic.
