+++ app/templates/projects.html
@@ -78,7 +78,7 @@
                     <i class="fas fa-folder-open text-primary"></i>
                 </div>
                 <h5 class="project-card-title">Open Existing Project</h5>
-                <p class="project-card-lead">Load a project you already use in PRISM Studio and validate it only when needed.</p>
+                <p class="project-card-lead">Load a project you already use in PRISM Studio here, then run full checks in the Validator.</p>
                 <div class="project-card-tags" aria-hidden="true">
                     <span class="project-card-tag">Load only</span>
                     <span class="project-card-tag">Sets current project</span>
